% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegimens.R
\name{createRegimens}
\alias{createRegimens}
\title{Create an oncology drug regimen table in a `writeDatabaseSchema` database}
\usage{
createRegimens(
  connectionDetails,
  cdmDatabaseSchema,
  writeDatabaseSchema,
  cohortTable,
  rawEventTable,
  regimenTable,
  regimenIngredientTable,
  vocabularyTable,
  cancerConceptId = 4115276,
  dateLagInput = 30,
  generateVocabTable = FALSE,
  generateRawEvents = FALSE,
  keepSteroids = FALSE,
  useHemoncToPullDrugs = FALSE
)
}
\arguments{
\item{keepSteroids}{Boolean parameter if TRUE algorithm will look for steroids along other drugs}
}
\value{
This function does not return a value. It is called for its side effect of
creating a new SQL table called `regimenIngredientTable` in `writeDatabaseSchema`.
}
\description{
Creates treatment regimens
}
